feat: implemented dashboard module with summary and monthly trend graph

+++ src/main/java/com/aiexpensetracker/dashboard/dto/response/CategorySpendingResponse.java
@@ -13,6 +13,6 @@ public record CategorySpendingResponse(
 
         BigDecimal amount,
 
-        Double percentage
+        BigDecimal percentage
 
 ) {}
